CBD-6267: Resovle issues with container health checks in AWS
In AWS, /home/couchbase and /home/couchbase/jenkins are on different
underlying volumes. This change ensures we carry out free space
checks on the overlayfs volume *and* the mounted workspace volume.
It also fixes a conditional statement which was checking for the
existence of the wrong file before searching for memory issue
indicators, and ensures we are cleaning up all unused images when
triggering `docker prune`.